Convert to typescript the slash commands create files
