--- app/slashcommands-create/client/client.js
@@ -1,7 +1,0 @@
-import { slashCommands } from '../../utils';
-
-slashCommands.add('create', null, {
-	description: 'Create_A_New_Channel',
-	params: '#channel',
-	permission: ['create-c', 'create-p'],
-});
